+++ R/readtextgrid.R
@@ -2,8 +2,12 @@
 #' Read a textgrid file into a tibble
 #'
 #' @rdname read_textgrid
-#' @param path a path to a dataframe
+#' @param path a path to a textgrid
 #' @param lines alternatively, the lines of a textgrid file
+#' @param encoding the encoding of the textgrid. The default value `NULL` uses
+#'   [readr::guess_encoding()] to guess the encoding of the textgrid. If an
+#'   encoding is provided, it is forwarded to `[readr::locale()]` and
+#'   `[readr::read_lines()]`.
 #' @param file an optional value to use for the `file` column. For
 #'   `read_textgrid()`, the default is the base filename of the input file. For
 #'   `read_textgrid_lines()`, the default is `NA`.
@@ -12,13 +16,18 @@
 #' @examples
 #' tg <- system.file("Mary_John_bell.TextGrid", package = "readtextgrid")
 #' read_textgrid(tg)
-read_textgrid <- function(path, file = NULL) {
+read_textgrid <- function(path, file = NULL, encoding = NULL) {
   if (is.null(file)) {
     file <- basename(path)
   }
 
-  path %>%
-    readr::read_lines() %>%
+  if (is.null(encoding)) {
+    encoding <- readr::guess_encoding(path)$encoding[1]
+  }
+  file_locale <- readr::locale(encoding = encoding)
+
+  path |>
+    readr::read_lines(locale = file_locale) |>
     read_textgrid_lines(file = file)
 }
 
@@ -31,35 +40,42 @@ read_textgrid_lines <- function(lines, file = NULL) {
 
   stopifnot(str_detect_any(lines, "ooTextFile"))
 
-  lines %>%
-    parse_textgrid_lines() %>%
-    tibble::as_tibble() %>%
+  lines |>
+    parse_textgrid_lines() |>
+    tibble::as_tibble() |>
     tibble::add_column(file = file, .before = 1)
 }
 
 
-#' Locate the path of the example textgrid file
+#' Locate the path of an example textgrid file
 #'
-#' Locate the path of the example textgrid file
+#' Locate the path of an example textgrid file
 #'
+#' @param which index of the textgrid to load
 #' @return Path of `"Mary_John_bell.TextGrid"` bundled with the `readtextgrid`
 #'   package.
 #'
 #' @details This function is a wrapper over [`system.file()`]  to locate the
-#' path to `"Mary_John_bell.TextGrid"`. This file is the default textgrid that
-#' is created by Praat.
+#' paths to bundled textgrids. These files are used to test or demonstrate
+#' functionality of the package.
+#'
+#' Two files are included:
+#'
+#' 1. `"Mary_John_bell.TextGrid"` - the default TextGrid created by Praat's
+#'    Create TextGrid command. This file is saved as UTF-8 encoding.
+#' 2. `"utf_16_be.TextGrid"` - a TextGrid with some IPA characters entered using
+#'    Praat's IPA character selector. This file is saved with UTF-16 encoding.
 #'
 #' @export
-example_textgrid <- function() {
-  system.file("Mary_John_bell.TextGrid", package = "readtextgrid")
+example_textgrid <- function(which = 1) {
+  basename <- c("Mary_John_bell.TextGrid", "utf_16_be.TextGrid")[which]
+  system.file(basename, package = "readtextgrid")
 }
 
-
-
 parse_textgrid_lines <- function(lines) {
-  lines %>%
-    slice_sections("item") %>%
-    purrr::map(parse_item_lines) %>%
+  lines |>
+    slice_sections("item") |>
+    purrr::map(parse_item_lines) |>
     plyr::ldply(as.data.frame, stringsAsFactors = FALSE)
 }
 
@@ -70,10 +86,9 @@ slice_sections <- function(lines, section_head) {
   purrr::map2(starts, ends, function(x, y) lines[seq(x, y, by = 1)])
 }
 
-
 parse_item_lines <- function(lines_items) {
-  item_num <- lines_items[1] %>%
-    stringr::str_extract("\\d+") %>%
+  item_num <- lines_items[1] |>
+    stringr::str_extract("\\d+") |>
     as.numeric()
 
   tier_type <- get_field(lines_items, "class")
@@ -104,10 +119,10 @@ parse_item_lines <- function(lines_items) {
 }
 
 parse_interval_tier <- function(lines_interval_tier) {
-  lines_interval_tier %>%
-    slice_sections("intervals") %>%
-    purrr::map(get_field_list, fields = c("xmin", "xmax", "text")) %>%
-    purrr::imap(add_annotation_num) %>%
+  lines_interval_tier |>
+    slice_sections("intervals") |>
+    purrr::map(get_field_list, fields = c("xmin", "xmax", "text")) |>
+    purrr::imap(add_annotation_num) |>
     plyr::ldply(as.data.frame, stringsAsFactors = FALSE)
 }
 
@@ -115,10 +130,10 @@ parse_point_tier <- function(lines_point_tier) {
   no_points <- str_detect_any(lines_point_tier, "points: size = 0")
 
   if (!no_points) {
-    df <- lines_point_tier %>%
-      slice_sections("points") %>%
-      purrr::map(get_field_list, fields = c("number", "mark")) %>%
-      purrr::imap(add_annotation_num) %>%
+    df <- lines_point_tier |>
+      slice_sections("points") |>
+      purrr::map(get_field_list, fields = c("number", "mark")) |>
+      purrr::imap(add_annotation_num) |>
       plyr::ldply(as.data.frame, stringsAsFactors = FALSE)
 
     # We treat points as zero-width intervals
@@ -157,11 +172,11 @@ get_field_list <- function(lines, fields) {
 get_field <- function(lines, field) {
   re <- paste0("(?<=", field, " = ).+")
 
-  lines %>%
-    stringr::str_extract(re) %>%
-    remove_na() %>%
-    utils::head(1) %>%
-    stringr::str_trim() %>%
+  lines |>
+    stringr::str_extract(re) |>
+    remove_na() |>
+    utils::head(1) |>
+    stringr::str_trim() |>
     str_unquote()
 }