+++ NEWS.md
@@ -1,3 +1,10 @@
+# readtextgrid dev version
+
+* Add `encoding` argument to `read_textgrid()`. (#7, #8, #9, @jonorthwash)
+* Include example UTF-16 textgrid for testing.
+* Remove magrittr dependency. (#5, @JoFrhwld)
+
+
 # readtextgrid 0.1.1
 
 * Fixes for CRAN resubmission.